Feat: Complete import validation and add search interface
This commit addresses several critical issues in the application, completes the import validation pipeline, and adds a new transaction search and edit feature.

**Part 1: Import Validation and Bug Fixes**

- **Fixed `CategoryAwareCSVImporter.js`:**
  - Patched a critical bug that caused an error on transactions with no description.
  - Improved the logic in `correctAccountAssignment` to more reliably identify "Tech Business" income from "PackerThomas".
  - Replaced the `processTransactions` method with a more robust, async version that validates each transaction, normalizes data, and logs any errors to a dedicated Firestore collection (`import_errors`).
  - Added a `normalizeDate` helper to ensure all dates are in `YYYY-MM-DD` format.

- **Fixed `UIManager.js`:**
  - Corrected the import process so that the import modal now closes automatically after a successful import.

**Part 2: Search & Edit Interface**

- **Created `TransactionSearch.js`:**
  - Added a new `TransactionSearch` class to manage the UI and logic for searching and filtering transactions.
  - The UI includes filters for search term, account, entity, and date range.

- **Integrated Search into `UIManager.js`:**
  - Added a "Search" button to the main application header.
  - Implemented `showSearchInterface` and `showDashboard` methods to toggle between the dashboard and the new search view.
  - The search results allow for quick inline category changes.

**Part 3: Data Recovery**

- **Added Data Recovery Function:**
  - Implemented a `recoverTechBusinessData` function in the main `app.js`.
  - This function scans existing transactions and retroactively fixes any that were miscategorized as not belonging to "Tech Business".
  - The function is exposed as `window.recoverTech()` for manual execution from the developer console.
